requirements.txt
discord.py==2.3.2
lavalink
aiohttp
async-timeout
PyNaCl